requirements.txt
Django==4.0.3
djangorestframework==3.13.1
django-environ==0.8.1
gunicorn==20.1.0
celery==5.2.6
psycopg2-binary==2.9.3
redis==4.2.2